Fix CI: Add --lib flag to release build in main-pr.yml
The release build step was missing --lib flag, causing it to build
examples which fail due to mutually exclusive features in feagi-async.

Using --lib is correct for crates.io publication as examples are not
included in published crates by default.

Examples are known to have API migration issues and will be fixed
separately.